refactor: utilize blueprints for route modularization
